src/middleware/monitoring.py
import time
from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware
from starlette.types import ASGIApp

from src.utils.monitoring import metrics_tracker

class MonitoringMiddleware(BaseHTTPMiddleware):
    def __init__(self, app: ASGIApp):
        super().__init__(app)
    
    async def dispatch(self, request: Request, call_next):
        # Start timer
        start_time = time.time()
        
        # Process request
        response = await call_next(request)
        
        # Calculate processing time
        processing_time = time.time() - start_time
        
        # Track request
        metrics_tracker.track_request(
            endpoint=request.url.path,
            method=request.method,
            status_code=response.status_code,
            processing_time=processing_time
        )
        
        return response 