requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
alembic>=1.7.1

# API and HTTP
aiohttp>=3.8.0
requests>=2.26.0
httpx>=0.23.0

# Voice processing
google-cloud-speech>=2.11.0
google-cloud-storage>=1.42.0
twilio>=7.8.0
pydub>=0.25.1

# AI and ML
openai>=1.0.0

# Utilities
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.0.1
structlog>=21.1.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
httpx>=0.23.0

# Development
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
isort>=5.9.3

# Notifications
sendgrid==6.8.0

# Speech Recognition and NLP
google-cloud-language==2.4.1
SpeechRecognition==3.8.1

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
aiohttp==3.9.1
uuid==1.30

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
jinja2==3.1.2

# Additional dependencies
redis==5.0.1
supabase==2.0.3 